Add option to selectively display user management errors
Added a option to display chosen error messages on the horizon page instead of
supressing them and displaying a general error message for user management in
identity dashboard.
This can be used to allow errors related to password policy violation to be
displayed. Any exception matching the whitelist will be shown.

This change adds a new option ERROR_WHITELIST in local_settings.py. Any error
messages included in this list will be whitlisted and displayed on horizon.
The list can also  include just a part of the error message.